feat: add tests ffor Api
